project/masking/main/requirements.txt
numpy 
torch
pillow
matplotlib
transformers
scipy
torchvision
unicodedata
opencv-python
scikit-image
math
os
collections





